--- a/src/node_external_reference.h
+++ b/src/node_external_reference.h
@@ -10,6 +10,9 @@
 
 namespace node {
 
+using CFunctionCallbackWithOneByteStringVoid =
+    void (*)(v8::Local<v8::Value>, const v8::FastOneByteString&);
+
 using CFunctionCallbackWithOneByteString =
     uint32_t (*)(v8::Local<v8::Value>, const v8::FastOneByteString&);
 using CFunctionCallback = void (*)(v8::Local<v8::Value> receiver);
@@ -96,6 +99,7 @@ class ExternalReferenceRegistry {
   V(CFunctionWithBool)                                                         \
   V(CFunctionBufferCopy)                                                       \
   V(CFunctionWriteString)                                                      \
+  V(CFunctionCallbackWithOneByteStringVoid)                                    \
   V(const v8::CFunctionInfo*)                                                  \
   V(v8::FunctionCallback)                                                      \
   V(v8::AccessorNameGetterCallback)                                            \