@tannal 2024-11-27
```js
// Flags: --expose-internals --no-warnings --allow-natives-syntax
'use strict';

const { internalBinding } = require('internal/test/binding');
const assert = require('assert');
const common = require('../common');

function testFastPaths() {
    const url = URL.createObjectURL(new Blob(['hello world', '🍌🍌🍌']));
    URL.revokeObjectURL(url);
}

if (common.isDebug) {
    const { getV8FastApiCallCount } = internalBinding('debug');

    eval('%PrepareFunctionForOptimization(URL.revokeObjectURL)');
    testFastPaths();
    eval('%OptimizeFunctionOnNextCall(URL.revokeObjectURL)');
    testFastPaths();

    assert.strictEqual(getV8FastApiCallCount('blob.revokeObjectURL'), 1);
}
```
This test always failed. The TurboFan didn't call the fast method.
But the fast method did get called while running the benchmark code above (the code in https://github.com/nodejs/node/pull/54712#issue-2499944222)
Am I missing something?